+++ src/main/java/com/sprint/hrbank/application/changelog/required/ChangeLogRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChangeLogRepository extends JpaRepository<ChangeLog, Integer> {}
+public interface ChangeLogRepository extends JpaRepository<ChangeLog, Long> {}
