+++ src/main/java/com/sprint/hrbank/application/employee/dto/EmployeeUpdateRequest.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 public record EmployeeUpdateRequest(
     String name,
     String email,
-    Integer departmentId,
+    Long departmentId,
     String position,
     LocalDate hireDate,
     EmployeeStatus status,
